Message: destructure styles & types
