+++ src/Message/types.ts
@@ -1,0 +1,26 @@
+import { ViewStyle, LayoutChangeEvent } from 'react-native'
+import { AvatarProps } from '../Avatar'
+import { SystemMessageProps } from '../SystemMessage'
+import { DayProps } from '../Day'
+import { IMessage, User, LeftRightStyle } from '../Models'
+import Bubble from '../Bubble'
+
+export interface MessageProps<TMessage extends IMessage> {
+  showUserAvatar?: boolean
+  position: 'left' | 'right'
+  currentMessage: TMessage
+  nextMessage?: TMessage
+  previousMessage?: TMessage
+  user: User
+  inverted?: boolean
+  containerStyle?: LeftRightStyle<ViewStyle>
+  renderBubble?(props: Bubble['props']): React.ReactNode
+  renderDay?(props: DayProps<TMessage>): React.ReactNode
+  renderSystemMessage?(props: SystemMessageProps<TMessage>): React.ReactNode
+  renderAvatar?(props: AvatarProps<TMessage>): React.ReactNode
+  shouldUpdateMessage?(
+    props: MessageProps<IMessage>,
+    nextProps: MessageProps<IMessage>,
+  ): boolean
+  onMessageLayout?(event: LayoutChangeEvent): void
+}
